image/book/requirement.txt
asgiref==3.8.1
Django==5.1.6
django-jinja==2.11.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-barcode==0.15.1
sqlparse==0.5.3
tzdata==2025.1
